pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "file-sanitizer"
version = "0.1.0"
description = "Upload files, sanitize contents (metadata/macros/scripts), and re-export safely with logs"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
  "Pillow==10.4.0",
  "pypdf==4.3.1",
]

[project.scripts]
file-sanitizer = "file_sanitizer.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
mypy_path = "src"
